Added a SparseBitVector class to the Vector hierarchy, using an IntSet. Intended to represent BitVectors / BitSets in which few of the bits will be populated (e.g. log-linear feature sets).
Renamed existing BitVector class as PackedBitVector and added a BitVector interface representing the methods shared between PackedBitVector and SparseBitVector.

We may eventually also want an ArrayBitVector or ImmutableBitVector class which would represent the same data structure, but use an int array for faster access (immutability would be an advantage since inserting into an array is expensive).