Aside from some minor renaming changes in various small parts of the code, the major change of this commit is the following:
In previous commits Aaron Dunlop had introduced a LogLinearMappedSequence class. This
LogLinearMappedSequence was breaking my very simple character alignment for morphology
that used a MultipleVocabularyMappedSequence. In particular Aaron had changed the method:

FullColumnAligner.backtraceWithGaps(
  MappedSequence sequence,
  ColumnAlignmentModel model,
  byte[][] backPointer,
  Vector gapVector,
  final float score)

so as to return a SequenceAlignment built from a ** LogLinearMappedSequence ** instead
of returning a SequenceAlignment built from a ** MultipleVocabularyMappedSequence **.
This commit changes the backtraceWithGaps method back to returning a SequenceAlignment
built from a MultipleVocabularyMappedSequence. If this breaks something that you need
Aaron, let me know and we can talk about what to do.