A batch of accumulated improvements to command-line tool infrastructure.
Fixed parsing of ambiguous date strings
Added support for listing required annotations in option annotations, controlling when the option is enabled.
A few special-case checks for use by Scala tools.
Added some convenience methods for commonly-used date values (useful for initializing default values of options).
Fixed a bug in handling a multi-valued option stored in an array with a default value (the default should not be included in the final array if a value is specified on the command-line).
Improved usage output.