Moved grammar init() method back into the constructor. Indecisive, I know. It makes for a longer constructor, but allows more local variables to be declared final, so the JIT can inline them.
Ignored non-terminals other than the start symbol which never occur as the child of a production. The Berkeley grammar includes a number of them.

Interned a lot of Strings in BaseSortedGrammar to save memory during the grammar pre-processing.

Handled unaries with sparse-matrix-vector multiplication in CsrSparseMatrixVectorParser. Speed impact is minimal, but it'll help us to move everything to a GPU and avoid main RAM -> GRAM traffic when we get to that point.

Handled gzipped grammar files and added support for CsrSparseMatrixVectorParser in ParserDriver