Changed visitCell() in ExhaustiveChartParser to take a start and end instead of a cell object; this will allow parsers more flexibility in how they represent charts (i.e., how object-oriented they want to be).
Added CellwiseExhaustiveChartParser, which implements the old visitCell(cell) signature, allowing most subclasses to remain unchanged.

Changed CellSelector to return a short[] instead of a ChartCell, to comply with the new visitCell() signature.

Annotated unit tests for the longer sentences with @DetailedTest and added a debug configuration that will test all parsers but only run the basic tests.