Genericized LinewiseCommandlineTool to accept task results other than Strings. Updated ParserDriver tasks to return ParseResult (nee ParseStats) instead of String, so that output of the stats and parse tree are atomic. This eliminates the jumbled out-of-order stats output.
Documented CellSelector.hasNext(), noting the requirement that parsers
call it until it returns false. Added a 'waitForActiveTasks()' method
to Parser, and called that at the end of each row in
LeftRightBottomTopTraversal.hasNext(). This allows a parser
implementation to spawn multiple tasks while parsing a row and wait on
a barrier before proceeding to dependent cells.

Implemented a row-parallel CSC which parallelizes CSC over each row,
using the aforementioned mechanism.