Renamed a few of the timing fields we maintain in ParseTask and reorganized the output at the end of each sentence.
Switched a couple of the finer-grained timings to accumulate in
nanoseconds (the total is still reported in milliseconds). Reported
total sentence time in milliseconds instead of seconds. We may have to
switch to microseconds someday, but that will be a good problem to
have...

Removed a few redundant timing fields.

Corrected reporting of chart init time, which was including FOM
initialization as well.